one last update to the gui 0 wind speed warning flags. for issue #442, related to issue #526
Noticed that the last fix wasn't quite handling more than one run in the gui correctly, and in the process found that the way the number of runs was counted when zeros in between rows needed updated, as well as that the wind speed table was only updating the warnings if the first row was updated. Was a bit worried to connect each of the 64 nRuns, but it didn't seem to slow anything down so I went ahead with the fix to allow proper nRun counting and updating at any changes to any row of wind speed and wind direction values.

Didn't see anything related to checking the values in the table if anything other than wind speed and wind direction were changed though. For now this should be good enough.