temporary fix to at least avoid seg faults for mass mesh sampling if the mass mesh sampling fails to produce a dataset.
Basically just avoid writing the outputs that require the dataset instead of referencing outside of an array, let it continue on to finish whatever else WindNinja is trying to output.

This occurs for two of the old openfoam 8 complex terrain smoke jumper cases, bruce and bear. Looks like the problem is that some of the points or faces aren't well enough defined for those meshes, and the openfoam sampling seg faults instead of skipping those sample points that trigger the problem. Seems to be caused by trying to break down faces or points for interpolation. I dealt with this with the old R script code, by manually running the sample to find the bad points/faces, then running code to find the x,y,z location of said points/faces, then remaking the sample point list dropping points from around that location and trying to sample again, repeating till it stopped finding bad points/faces in the list.