extended the old ascii2png kml bounds fix to vector kml output
the original fix was to deal with png layers not quite extending to the right spots/bounds in the domain. The fix was necessary this time because negative input min and max bounds would result in vectors not being drawn in google earth. The locations of the vectors and values would pop up, but no arrows would be drawn.

The problem this time turned out to be because the bounds in the overall LatLonAltBox of the kmz would get reversed. The fix worked because it keeps track of all four points of the bounding box, rather than just two. I suspect that the fix will also be more robust for when we allow more types of projections in WindNinja, as the bounding box can take on varying orientations during reprojections.

The original fix is described in e765438e5a0d655a791c0f699687947dd58ac3a9 and 395c93eb198f9b671232c449951c62ecae366c60. This fix should come in real handy for issue #544