re-enable simpleFoam return false for OpenFOAM 8 simulations
this was originally disabled in the openfoam 8 upgrade, as foam 8 simulations on ubuntu 16.04 would still run correctly, but would return a non-zero error code. Enabling this again allows simpleFoam to be retried upon failure to converge.

I also caught that the newly added --allow-run-as-root option in parallel calls to moveDynamicMesh and simpleFoam is not a legal option in openfoam 2.2.0

should just be me still trying to run openfoam 2.2.0 on ubuntu 16.04, for foam version testing purposes, otherwise we've moved on to newer openfoam and ubuntu.

related to issue #458