made the angleFromNortha signed value, for issue #544
This was required to get the angle correction done right for utm projections going from east to west, the original calculation turned out to be good for utm projections going from west to east. Basically, when reprojecting, the angle corrector needed to be added for datasets that rotate counter-clockwise, and subtracted for datasets that rotate clockwise.