more work on consistent color scheme for wind vectors in KMZ output, for issue #441
I tried fixing the KmlVector::calcSplitValsFromSplitVals() calculations, by using the min and the max for the first and the last values in the scale. And by using something that came from experimenting with the lever rule.

The lever rule scaling came out to be a consistent color scale for both equal_interval and equal_color scaling, but the scale does not come out to be the same as when just packing all the various ascii grids into one single ascii grid for the equal_color scaling, seems to bias towards higher sets of values as it seems to scale not just to the mean of a given dataset interval but also to the full range of the given dataset