first pass, PASTCAST /vsigs/ download as GTIFF format instead of standard GRIB (GRIB2) format, hopefully this lets it work for Windows builds, for #514
we were quite surprised that, just by changing the file extension/format, the resulting files were still readable and useable for full simulations by the code. Looks like that part is still working, because GDALTranslate just packed all the GRIB2 style metadata into the GTIFF without changing any of the metadata naming schemes. Still have to decide whether we try to make it more correct or not, or just use it as is.