major update, cleanup, and testing for dem no data filler utility script, for issue #261
fixed check for isLcp file, if GTIFF wasn't enough anymore, use if nBands > 1

added a user prompt, to pause and warn the user that the input_dem_file will be overwritten if NO_DATA values are found by the script, asking the user whether to continue or to exit the script without touching the input_dem_file. Added a small snippet to the Usage()/help() information to let the user know that this user prompt will occur

added option to choose whether to NO_DATA fill the remaining vegetation bands for lcp files, or just the 1st elevation band. Improved the Usage()/help() information to better describe the what/why to do NO_DATA filling of only the 1st elevation band, or of all the bands, vegetation bands as well as the 1st elevation band. Also added to the Usage()/help() information details about what the specific 8 bands are that are expected to be in lcp files

also updated the NO_DATA checking, to handle varying numbers of NO_DATA values between the various bands, so the script can potentially be rerun on an lcp file even if the script was already run on the lcp file to fill just the 1st elevation band. In the process, updated the script to be smarter about whether to trigger various data read-in, filling, and write processes to be more efficient if no NO_DATA values were found, or if only the 1st elevation band is to be worked on

various other cleanup, including of the various info and error/debug messages