updating angleFromNorth and prj1 to prj2 style calculations, for a new and more consolidated convention, but just for domainAverage runs and nomads weather model runs till we get through testing the convention. This was technically where I got the code to for testing our new convention, there are definitely problems with the results that imply that I am still off by a half or a whole sign somewhere in either the convention, or in my following of the convention. For issue #544.
The new convention, is that the coordinateTransformAngle is always defined as going from prj1 to prj2, that coordinateTransformAngle is positive clockwise and negative counter-clockwise, which came out to be that to go from prj1 to prj2, the formula is to always subtract the coordinateTransformAngle. So "prj2 = prj1 - coordinateTransformAngle" where "coordinateTransformAngle = coordinateTransformAngle_from_prj1_to_prj2". This means that "coordinateTransformAngle_from_prj2_to_prj1 = -coordinateTransformAngle_from_prj1_to_prj2".

This means a new convention for the angleFromNorth. Now, the convention is that the stored angleFromNorth value always needs to be going FROM geographic lat/lon coordinates TO dem projection coordinates. So "angleFromNorth = coordinateTransformAngle_from_N_to_dem". Because we store/calculate coordinateTransformAngle as coordinateTransformAngle_from_dem_to_N for this case, and "coordinateTransformAngle_from_N_to_dem = -coordinateTransformAngle_from_dem_to_N", this means that the stored angleFromNorth is equivalent to "angleFromNorth = -coordinateTransformAngle" when looking at the code.

In the past, the angleFromNorth was defined as going FROM dem projection coordinates TO geographic lat/lon coordinates, as a positive value, and so we used to add it to go from geographic lat/lon coordinates to dem projection coordinates, and used to subtract it to go from dem projection coordinates to geographic lat/lon coordinates. So exactly the reverse of all the signs of before, for both add/subtract operations, and for the stored angleFromNorth value. If I have this correct, and I am following the conventions correctly, then this means that "angleFromNorth_old = coordinateTransformAngle_from_dem_to_N = -coordinateTransformAngle_from_N_to_dem = -angleFromNorth_new". But it looks like something might be off in how I am applying this convention, and/or the kmz geographic lat/lon calculation comes out to be opposite in sign for its formula compared to before, I'm still trying to figure out what I have messed up. Hope this description isn't too confusing.