fix to last prj1 to prj2 style calculation, for issue #544
accidentally forgot the new rule, for the kmz output step, that the coordinateTransformAngle used for each coordinate transform, needs to always be the one calculated for the specific FROM prj1 TO prj2 or FROM prj2 TO prj1 respectively, no using the (-1)*coordinateTransformAngle of the opposite direction

but, following this rule resulted in the 270 degree flat_test case still having arrows going from west to east properly, but the printed values came out as ~262 instead of 270, so it is wrong. This implies that the above rule needs to be revised and tested some more, but I thought it best to update the code with this implementation to make the next testing a bit easier.