started to update additional ascii output (ascii_out_4326, ascii_out_json, ascii_out_uv, etc), for the angleFromNorth calculations. For issue #544, related to pull requests #468 and #471.
It should be working, but something weird is going on. The resulting data appear to be warped to the correct locations, the source data cells after warp are correct values, but the NO_DATA region that always gets added around the warped data is full of garbage values that don't match the NO_DATA value of the dataset.

That being said, I was able to replicate all of the previously added additional template AsciiGrid::write_Grid() type functions, by just calling AsciiGrid::write_Grid() and AsciiGrid::write_json_Grid(), so I removed the unused template AsciiGrid::write_Grid() type functions. I also paired down the ninja::writeAsciiOutputFiles() style functions down to a single function, which is now more correct even without the angleFromNorth corrections, because it warps the ascii grids to EPSG:4326 as u and v grids rather than spd and dir grids that suffer from angle interpolation problems in gdal (dir is a periodic value).