found an error where GDALAutoCreateWarpedVRT() was not properly setting NO_DATA values for cases where the source gdal dataset did not contain a set NO_DATA value. Also fixed an error were AsciiGrid::get_meanValue() wasn't properly ignoring the NO_DATA value when returning the mean. Related to issue #544.
When using GDALAutoCreateWarpedVRT() with a source gdal dataset that does not contain a set NO_DATA value, the normal pszWarpOptions for setting NO_DATA stuff ends up ignored, even setting INIT_DEST=NO_DATA does not work, the warped result ends up using 0.0 values in place of NO_DATA values. Trying to call GDALSetRasterNoDataValue() for every band pre-warp does not work, though it still lets the call occur, probably because the gdal dataset is opened as read only. Calling GDALSetRasterNoDataValue() for every band post-warp sets the value, but they are already initialized to a value of 0.0 by that point.

The only method that seemed to work, was setting INIT_DEST to the exact NO_DATA value. This fix is currently implemented for the additional output ascii grid (4326) and the griddedInitialization north up datasets, I tested the fix for nmads_wx_init.cpp and it WORKED, though I decided to wait to add it for now because it probably needs more testing.