fixed menu statusTips failing to display, by turning them into toolTips, and setting toolTipsVisible=true for each QMenu in QMenuBar. For issue #638
Apparently statusTips only show up if the QMainWindow has a QStatusBar, which we are purposefully not doing, not enough control with its location and style. An added QStatusBar just puts it at the bottom of the page without easy ways to edit it, which is NO LONGER displaying statusTips over the menu items anymore.

Note, even if toolTipsVisible is set to true, and QActions supposedly default the toolTip value to the text of the QAction, you have to explicitly set the toolTip values for each given QAction or they won't show up.