2nd attempt at adding the progress bar, for issue #654
I managed to accidentally get rid of the weird final extra flush-like emit signals, that were leading to the final message and close being replaced with another message. But, I am still unsure whether the method is behaving as a code guard, ignoring the still emitted problem signals, or, if the method is indirectly forcing the event loop to FINALLY properly wrap up the signals BEFORE the next set. It's really hard to tell, it's possible the problem comes back when scaling up to more threads.

Regardless, the method is working. The specific ordering of the code was VERY important. For a canceling solve(), needed to destroy the ninjaArmy first to try to force the emitted signals to wrap up before doing anything else. For both a canceling solve() and a regularly finishing solve(), the next step is to clean up the QMessageBox, by first running disconnect() of cancelSolve() to convert the "cancel" button into a "close" button, avoiding cancelSolve() being activated again when clicking the "close" button, then second updating the final QMessageBox text, message, and values. Then, because all these uses of the stored runProgress are finally wrapped up, the stored progress for each run is finally cleared, and the rest of the code can finally continue.