1st attempt at refactoring ninjaCom, to remove the need for the various ninjaComHandler() classes, for issue #666
succeeded in dropping the various ninjaComHandler() classes from ninjaCom, though every bit of the logic differences between the various ninjaComHandler() classes was also dropped. I also did my best to refactor as little as possible, in such a way that git diff wouldn't get confused when showing the changes, so there is still cleanup to do even with this commit. But, while the cli stdout is accidentally commented out, the gui is running fine for this commit.