tested and updated the build_deps_ubuntu_2404.sh build script, while attempting to compile OpenFOAM 8 and 9 from source on Ubuntu 24.04
bad luck with OpenFOAM 8 and 9 on Ubuntu 24.04, they compiled fine, but they had weird runtime IO errors that implied the need to use an older compiler. I found a workaround, described in the script, I went to the .deb package from the openfoam website, downloaded it, and unpacked it, that worked better so far as I can tell.

compiling gdal required "--with-unix-stdio-64=no", some kind of int64 vs int32 problem. Supposedly another workaround might be "sudo apt install g++-multilib gcc-multilib", but it still needs investigated and tested as a method.