more cleanup to ninjaCom, including enabling and testing a multiStream log file in the "test_capi_domain_average_wind.c" C-API auto-test function, and moving the set_ninjaCommunications() copy constructor/assignment operator methods to the ninja and ninjaArmy copy constructors and assignment operators.
Moving the set_ninjaCommunications() methods made it possible to completely remove the set_ninjaCommunication() calls in both ninjaArmy and ninja, as that step is now done via the ninja and ninjaArmy constructors/assignment operators. That being said, the runNumber for the simulation still had to be set alongside the ninjaCom runNumber, so a new function called set_ninjaComRunNumber() had to replace calls to set_ninjaCommunication(), for the individual ninjas[i] level ninjaCom. The set_ninjaComRunNumber() function was able to be completely dropped for the ninjaArmy level ninjaCom.

During testing of the multiStream log file methods, I discovered a few ninjas[i] level ninjaCom data members that weren't being properly set at the ninjaArmy level, so I also did a bit of cleanup to the various set_ninjaCom() related functions and methods within the C-API, ninjaArmy, and ninja. Effectively I was trying to make sure that the given functions/methods were being set or called at the appropriate level, and only at the appropriate level, made for a better organization of the methods.

For issue #666, related to issues #639 and #582.